requirements.txt
python-3.11.9
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
firebase-admin==6.2.0
cryptography==41.0.7

langchain>=0.1.0,<0.2.0
langchain-groq==0.1.0

google-generativeai==0.3.0
redis==5.0.1